Move dataset lookup to separate module, use template
