+++ src/views/license.j2
@@ -1,0 +1,6 @@
+<p><b>License:</b> {{ license }}</p>
+<p><b>Make:</b> {{ make }}</p>
+<p><b>Model:</b> {{ model }}</p>
+<p><b>Department:</b> {{ dept }}</p>
+<p><b>Description:</b> {{ descrip }}</p>
+<p><b>Equipment Type:</b> {{ equipment_type }}</p>
